requirements.txt
python-telegram-bot==21.5
gspread
google-auth
matplotlib